:recycle: refactor: ajusta tabela de símbolos para seguir estrutura de pastas anterior
- Adiciona caso if em `ast_copy`
- Move função `main` do parser para o arquivo `main.c`
- Remove parâmetro `var`
