+++ src/scanner.l
@@ -45,7 +45,6 @@ static char *unescape_string(const char *src) {
 
 %%
 
-"var"                   { return VAR; }
 "true"                  { yylval.boolValue = 1; ADVANCE_COLUMN; return BOOL_LIT; } /* booleanos */
 "false"                 { yylval.boolValue = 0; ADVANCE_COLUMN; return BOOL_LIT; }
 
@@ -56,33 +55,34 @@ static char *unescape_string(const char *src) {
                           // Processa string com escapes
                           char *str = (char*)malloc(yyleng); // Aloca espaço suficiente
                           int str_index = 0;
-                          int text_index = 1; // Pula a aspas inicial
+                          size_t text_index = 1; // Pula a aspas inicial
+                          size_t end = (size_t)yyleng - 1; // Posição da aspas final
 
-                          while (text_index < yyleng - 1) { // Para antes da aspas final
-                            if (yytext[text_index] == '\\') {
-                              // Sequência de escape
-                              text_index++;
-                              switch (yytext[text_index]) {
-                                case 'n': str[str_index++] = '\n'; break;
-                                case 't': str[str_index++] = '\t'; break;
-                                case 'r': str[str_index++] = '\r'; break;
-                                case 'b': str[str_index++] = '\b'; break;
-                                case 'f': str[str_index++] = '\f'; break;
-                                case 'v': str[str_index++] = '\v'; break;
-                                case '\\': str[str_index++] = '\\'; break;
-                                case '\"': str[str_index++] = '\"'; break;
-                                case '\'': str[str_index++] = '\''; break;
-                                case '0': str[str_index++] = '\0'; break;
-                                default:
-                                  // Se não for um escape reconhecido, mantém o caractere
-                                  str[str_index++] = '\\';
-                                  str[str_index++] = yytext[text_index];
-                                  break;
+                          while (text_index < end) { // Para antes da aspas final
+                              if (yytext[text_index] == '\\') {
+                                // Sequência de escape
+                                text_index++;
+                                switch (yytext[text_index]) {
+                                    case 'n': str[str_index++] = '\n'; break;
+                                    case 't': str[str_index++] = '\t'; break;
+                                    case 'r': str[str_index++] = '\r'; break;
+                                    case 'b': str[str_index++] = '\b'; break;
+                                    case 'f': str[str_index++] = '\f'; break;
+                                    case 'v': str[str_index++] = '\v'; break;
+                                    case '\\': str[str_index++] = '\\'; break;
+                                    case '\"': str[str_index++] = '\"'; break;
+                                    case '\'': str[str_index++] = '\''; break;
+                                    case '0': str[str_index++] = '\0'; break;
+                                    default:
+                                        // Se não for um escape reconhecido, mantém o caractere
+                                        str[str_index++] = '\\';
+                                        str[str_index++] = yytext[text_index];
+                                        break;
+                                }
+                              } else {
+                                str[str_index++] = yytext[text_index];
                               }
-                            } else {
-                              str[str_index++] = yytext[text_index];
-                            }
-                            text_index++;
+                              text_index++;
                           }
                           str[str_index] = '\0';
